+++ tubular/comparison.py
@@ -46,6 +46,9 @@ class EqualityChecker(
     lazyframe_compatible: bool
         class attribute, indicates whether transformer works with lazyframes
 
+    deprecated: bool
+        indicates if class has been deprecated
+
     """
 
     polars_compatible = False
@@ -55,6 +58,8 @@ class EqualityChecker(
     FITS = False
 
     jsonable = False
+
+    deprecated = True
 
     @beartype
     def __init__(