+++ tubular/imputers.py
@@ -1188,6 +1188,9 @@ class NearestMeanResponseImputer(BaseImputer):
     lazyframe_compatible: bool
         class attribute, indicates whether transformer works with lazyframes
 
+    deprecated: bool
+        indicates if class has been deprecated
+
     """
 
     polars_compatible = True
@@ -1197,6 +1200,8 @@ class NearestMeanResponseImputer(BaseImputer):
     jsonable = False
 
     FITS = True
+
+    deprecated = True
 
     def __init__(
         self,