+++ tubular/nominal.py
@@ -1964,6 +1964,9 @@ class OrdinalEncoderTransformer(
     lazyframe_compatible: bool
         class attribute, indicates whether transformer works with lazyframes
 
+    deprecated: bool
+        indicates if class has been deprecated
+
     """
 
     polars_compatible = False
@@ -1973,6 +1976,8 @@ class OrdinalEncoderTransformer(
     jsonable = False
 
     FITS = True
+
+    deprecated = True
 
     def __init__(
         self,
@@ -2154,6 +2159,9 @@ class NominalToIntegerTransformer(BaseNominalTransformer, BaseMappingTransformMi
     lazyframe_compatible: bool
         class attribute, indicates whether transformer works with lazyframes
 
+    deprecated: bool
+        indicates if class has been deprecated
+
     """
 
     polars_compatible = False
@@ -2163,6 +2171,8 @@ class NominalToIntegerTransformer(BaseNominalTransformer, BaseMappingTransformMi
     jsonable = False
 
     FITS = True
+
+    deprecated = True
 
     def __init__(
         self,